twin/pyproject.toml
[project]
name = "auxin-twin"
version = "0.1.0"
description = "PyBullet digital twin — TwinSource, scene, trajectory, renderer, websocket server"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "Apache-2.0" }
authors = [
    { name = "Edwin Redhead" },
    { name = "Tara Kasayapanand" },
]
dependencies = [
    "pybullet>=3.2.6",
    "numpy>=1.26.0",
    "anyio>=4.0.0",
    "websockets>=12.0",
    "imageio>=2.34.0",
    "imageio-ffmpeg>=0.5.1",
    "pillow>=10.0.0",
    # auxin-sdk consumed as a path dep so TwinSource can subclass TelemetrySource
    "auxin-sdk",
]

[project.scripts]
twin = "twin.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.4.0",
]

[tool.uv.sources]
auxin-sdk = { path = "../sdk", editable = true }

[tool.hatch.build.targets.wheel]
packages = ["src/twin"]

[tool.ruff]
target-version = "py311"
line-length = 100
src = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--tb=short"
